navigation not as Parameter on AppTopBar and renaming Impressum to ImpressumScreen
